requirements.txt
torch
torchaudio
torchvision
duckietown-gym-daffy
numpy
gym
opencv-python
PyYAML
matplotlib
scipy
seaborn
simple_pid
numba
pyglet
Pillow